Fixes after static analysis to look for bugs in Android code (#64)
* Unused code

* Redundant null check

* Performance: Fixed String concatenation inside a loop with +

* Performance: Innefficient map interations

* Performance: Unbox/rebox not neccesary

* Performance: Boxing/unboxing to parse a primitive

* Performance: static fields

* Graddle updated

* Fix for shadow the simple name of the superclas

SeekBar shadows the simple name of the superclass android.widget.SeekBar

* SeekBar shadows the simple name of the superclass android.widget.SeekBar

* Removed WARNING: Dependency org.json:json:20131010 is ignored for debug

* Field Show.imdbId masks field in superclass butter.droid.base.providers.media.models.Media

* Unread field type

* Fixed incorrect comparision type

* Fix for Malicious code vulnerability (Mutable static field)

* Fix exposing internal representation...
by incorporating reference to mutable object

* Fixed possible null pointer dereference.

If is not a movie it is a show, we always initialize the media object as a movie.

* Fixed guaranteed dereferenced Null value

* Cleanup streams

* Static inner classes

This classes are an inner classes and does not use its embedded reference to the object which created it. This reference makes the instances of the class larger, and may keep the reference to the creator object alive longer than necessary.

* Unused fields

* Clean up streams

* Clean up streams

* Bad practice, missing comparative methods.

* Fixed redundant Null check

* Fixed possible null pointer dereference

+++ base/src/main/java/butter/droid/base/content/preferences/PreferencesHandler.java
@@ -73,7 +73,7 @@ public interface PreferencesHandler {
                     .setOnClickListener(new PrefItem.OnClickListener() {
                         @Override
                         public void onClick(final PrefItem item) {
-                            handler.openListSelection(item.getTitle(), items, SelectionMode.SIMPLE_CHOICE, (int) item.getValue(), 0, 0, new OnSelectionListener() {
+                            handler.openListSelection(item.getTitle(), items, SelectionMode.SIMPLE_CHOICE, item.getValue(), 0, 0, new OnSelectionListener() {
                                 @Override
                                 public void onSelection(int position, Object value) {
                                     item.saveValue(position);
